requirements.txt.txt
streamlit
python-docx
pandas
pillow
requests
PyMuPDF
